удаляем PlaylistController и PlaylistService, пишем для Customer, Invoice, InvoiceLine service/controller layer, дорабатываем вывод ошибки для всех контроллеров если id не найден

+++ src/main/java/com/personal/Chinook/services/entity_services/AlbumService.java
@@ -2,10 +2,13 @@ package com.personal.Chinook.services.entity_services;
 
 import com.personal.Chinook.DTO.AlbumDTO;
 import com.personal.Chinook.DTO.AlbumSaveDto;
+import com.personal.Chinook.exceptions.custom.InvalidFieldException;
 import com.personal.Chinook.exceptions.custom.NotFoundInDBException;
 import com.personal.Chinook.mapper.AlbumMapper;
 import com.personal.Chinook.models.Album;
+import com.personal.Chinook.models.Track;
 import com.personal.Chinook.repositories.AlbumRepository;
+import com.personal.Chinook.repositories.TrackRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,11 +21,12 @@ import java.util.UUID;
 public class AlbumService {
 
     private final AlbumRepository albumRepository;
+    private final TrackRepository trackRepository;
     private final AlbumMapper albumMapper;
 
     @Transactional(readOnly = true)
     public AlbumDTO getAlbumById(UUID id) throws NotFoundInDBException {
-        Album album = albumRepository.findById(id).orElseThrow(() -> new NotFoundInDBException(""));
+        Album album = albumRepository.findById(id).orElseThrow(() -> new NotFoundInDBException("Не найден Альбом по id = "+ id));
         return albumMapper.toAlbumDTO(album);
     }
 
@@ -43,7 +47,7 @@ public class AlbumService {
 
     @Transactional
     public AlbumDTO updateAlbum(AlbumDTO albumDTO) throws NotFoundInDBException {
-        Album albumEntity = albumRepository.findById(albumDTO.getId()).orElseThrow(() -> new NotFoundInDBException(""));
+        Album albumEntity = albumRepository.findById(albumDTO.getId()).orElseThrow(() -> new NotFoundInDBException("Не найден Альбом по id = "+ albumDTO.getId()));
         if (albumMapper.toAlbumDTO(albumEntity).equals(albumDTO)) {
             return albumMapper.toAlbumDTO(albumEntity);
         }
@@ -54,9 +58,14 @@ public class AlbumService {
 
     @Transactional
     public AlbumDTO deleteAlbumById(UUID id) throws NotFoundInDBException {
-        Album artist = albumRepository.findById(id).orElseThrow(() -> new NotFoundInDBException("asrd"));
+        Album album = albumRepository.findById(id).orElseThrow(() -> new NotFoundInDBException("Не найден Альбом по id = "+ id));
+        List<Track> tracks = trackRepository.searchByAlbumId(id);
+        if(!tracks.isEmpty()) {
+            throw new InvalidFieldException("Перед тем как удалить Альбом c id ["+ id +"], необходимо удалить связанные с ним треки "
+                    +"\n" + tracks.stream().map(Track::getId).toList());
+        }
         albumRepository.deleteById(id);
-        return albumMapper.toAlbumDTO(artist);
+        return albumMapper.toAlbumDTO(album);
     }
 
     @Transactional(readOnly = true)
@@ -68,4 +77,14 @@ public class AlbumService {
         return albumMapper.toAlbumDTOs(albums);
     }
 
+//    @Transactional
+//    public AlbumDTO deleteAlbumWithTracksById(UUID id) throws NotFoundInDBException {
+//        Album album = albumRepository.findById(id).orElseThrow(() -> new NotFoundInDBException("asrd"));
+//        List<Track> tracks = trackRepository.searchByAlbumId(id);
+//        if(!tracks.isEmpty()) {
+//
+//        }
+//        albumRepository.deleteById(id);
+//        return albumMapper.toAlbumDTO(album);
+//    }
 }
